requirements.txt
pytesseract
pillow
pymupdf
pdf2image
mysql-connector-python
python-dotenv
requests
streamlit